Create Lab4-Assignment4.asm
To dectect overflow in additional operation, we also use other rule than the one
in Assignment 1. This rule is: when add two operands that have the same sign,
overflow will occur if the sum doesn’t have the same sign with either operands.
You need to use this rule to write another overflow detection program.